+++ ethereum/p2p/src/main/java/org/hyperledger/besu/ethereum/p2p/rlpx/RlpxAgent.java
@@ -35,6 +35,7 @@ import org.hyperledger.besu.ethereum.p2p.rlpx.wire.PeerConnectionGatekeeper;
 import org.hyperledger.besu.ethereum.p2p.rlpx.wire.messages.DisconnectMessage.DisconnectReason;
 import org.hyperledger.besu.plugin.data.EnodeURL;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
+import org.hyperledger.besu.util.CacheMaintenanceExecutor;
 import org.hyperledger.besu.util.Subscribers;
 
 import java.net.InetSocketAddress;
@@ -43,15 +44,14 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 import jakarta.validation.constraints.NotNull;
 import org.apache.tuweni.bytes.Bytes;
 import org.slf4j.Logger;
@@ -74,10 +74,10 @@ public class RlpxAgent {
   private final Supplier<Stream<PeerConnection>> allConnectionsSupplier;
   private final Supplier<Stream<PeerConnection>> allActiveConnectionsSupplier;
   private final Cache<Bytes, CompletableFuture<PeerConnection>> peersConnectingCache =
-      CacheBuilder.newBuilder()
+      Caffeine.newBuilder()
           .expireAfterWrite(
               Duration.ofSeconds(30L)) // we will at most try to connect every 30 seconds
-          .concurrencyLevel(1)
+          .executor(CacheMaintenanceExecutor.getInstance())
           .build();
 
   private RlpxAgent(
@@ -235,14 +235,10 @@ public class RlpxAgent {
     final Optional<DisconnectReason> maybeDisconnectReason = gatePeerConnection(peer, false);
 
     if (maybeDisconnectReason.isEmpty()) {
-      try {
-        synchronized (this) {
-          peerConnectionCompletableFuture =
-              peersConnectingCache.get(
-                  peer.getId(), () -> createPeerConnectionCompletableFuture(peer));
-        }
-      } catch (final ExecutionException e) {
-        throw new RuntimeException(e);
+      synchronized (this) {
+        peerConnectionCompletableFuture =
+            peersConnectingCache.get(
+                peer.getId(), ignored -> createPeerConnectionCompletableFuture(peer));
       }
     } else {
       final String errorMsg =